requirements.txt
opencv-python
numpy
ai-edge-litert
pillow